requirements.txt
kivy
google-genai
python-dotenv